PPLUS-749: Upgrader creates PR in GitLab / Renamed some models
